Refactoring: Bundling with Webpack, automated testing
This commit refactors the automated tests.
It solves configuration and imports, but some of the tests
are now failing, we need to analyse now the failures and fix
them.
