+++ docs/adding-webpack-notes.md
@@ -6,3 +6,7 @@
 - https://www.timroes.de/using-ecmascript-6-es6-with-angularjs-1-x
 - https://stackoverflow.com/questions/28969861/managing-jquery-plugin-dependency-in-webpack
 - https://github.com/monospaced/angular-qrcode#es2015--webpack
+- Testing
+	- https://mike-ward.net/2015/09/07/tips-on-setting-up-karma-testing-with-webpack/
+	- https://www.npmjs.com/package/karma-webpack
+	- https://gist.github.com/trinitroglycerin/68754b920df83a977f12
