+++ package.json
@@ -18,6 +18,7 @@
     "ng-infinite-scroll": "^1.3.0"
   },
   "devDependencies": {
+    "angular-mocks": "1.4.7",
     "bower": "^1.8.8",
     "chrome-webstore-upload-cli": "^1.1.1",
     "clean-webpack-plugin": "^4.0.0-alpha.0",
@@ -29,10 +30,12 @@
     "karma": "^4.4.1",
     "karma-chrome-launcher": "^2.2.0",
     "karma-jasmine": "^1.1.1",
+    "karma-webpack": "^5.0.0",
     "mini-css-extract-plugin": "^1.6.0",
     "url-loader": "^4.1.1",
     "webpack": "^5.39.1",
-    "webpack-cli": "^4.7.2"
+    "webpack-cli": "^4.7.2",
+    "webpack-merge": "^5.8.0"
   },
   "scripts": {
     "start": "webpack --watch --config webpack.dev.js",
