+++ spec/background/playlist.spec.js
@@ -1,10 +1,17 @@
-'use strict';
+import $ from 'jquery';
+import podStationBackgrounAppModule from '../../src/background/ng/backgroundApp';
+import { ajaxGetFeed, ajaxGetFeedFromFile } from '../reuse/ajax.mock';
+import analyticsServiceMockFn from '../reuse/analyticsService.mock';
+import browserStorageMockFn from '../reuse/browser.mock';
+import fixAngularInjector from '../reuse/fixAngularInjector';
+import messageServiceMockFn from '../reuse/messageService.mock';
+import FEEDS from './feeds/feedsConstants';
 
 describe('playlist', function() {
 
-	beforeEach(module('podstationBackgroundApp'));
+	beforeEach(angular.mock.module(podStationBackgrounAppModule.name));
 
-	beforeEach(module(function($provide) {
+	beforeEach(angular.mock.module(function($provide) {
 		$provide.factory('browser', browserStorageMockFn);
 
 		// Dummies
@@ -19,7 +26,7 @@ describe('playlist', function() {
 	var podcastDataService;
 	var ajaxSpy;
 
-	beforeEach(inject(function($injector) {
+	beforeEach(angular.mock.inject(function($injector) {
 		fixAngularInjector($injector);
 
 		ajaxSpy = spyOn($, 'ajax').and.callFake(ajaxGetFeed);
