+++ spec/reuse/messageService.mock.js
@@ -6,3 +6,5 @@ function messageServiceMockFn() {
 		}
 	}
 }
+
+export default messageServiceMockFn;
