+++ src/background/entities/podcast.js
@@ -3,6 +3,7 @@
 import $ from 'jquery';
 import { getBrowserService, getMessageService } from "../../reuse/ng/reuse";
 import { getNotificationManagerService } from '../ng/serviceGetters';
+import parseEmailTag from '../utils/parseEmailTag';
 
 /**
  * 
@@ -488,27 +489,6 @@ function Podcast(url) {
 			});
 		}
 	}
-}
-
-/**
- * A very cheap email tag parser, does not support mailto links
- * e-mail tags do not have a well defined format :(
- * see 
- * @param {string} tagContent
- * @returns {string} returns a string containing an e-mail
- */
-function parseEmailTag(tagContent) {
-	let email = undefined;
-	
-	if(tagContent) {
-		let e = tagContent.split(' ');
-
-		if(e[0] !== 'noreply@blogger.com') {
-			email = e[0];
-		}
-	}
-	
-	return email;
 }
 
 export default Podcast;
